requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b9
aiogram-dialog==2.0.0b21
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
arq==0.26.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.7
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
hiredis==3.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.1.0
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
redis==5.0.8
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.32
StrEnum==0.4.15
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
watchdog==2.3.1
Werkzeug==2.3.7
WTForms==3.1.2
yarl==1.9.6
